# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: La Scarpetta Restaurant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gallery:
    get:
      operationId: listGalleryPhotos
      tags: [gallery]
      summary: List gallery photos
      parameters:
        - in: query
          name: type
          schema:
            type: ["string", "null"]
            enum: [dishes, interior, events, daily, null]
        - in: query
          name: limit
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of gallery photos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryPhoto"
    post:
      operationId: createGalleryPhoto
      tags: [admin, gallery]
      summary: Add a gallery photo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryPhotoInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryPhoto"
  /gallery/{id}:
    put:
      operationId: updateGalleryPhoto
      tags: [admin, gallery]
      summary: Update a gallery photo
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryPhotoInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryPhoto"
    delete:
      operationId: deleteGalleryPhoto
      tags: [admin, gallery]
      summary: Delete a gallery photo
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ──────────────────── EVENTS ────────────────────
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events
      parameters:
        - in: query
          name: upcoming
          schema:
            type: ["boolean", "null"]
        - in: query
          name: lang
          schema:
            type: string
            default: ru
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [admin, events]
      summary: Create an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    put:
      operationId: updateEvent
      tags: [admin, events]
      summary: Update an event
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [admin, events]
      summary: Delete an event
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ──────────────────── TABLE RESERVATIONS ────────────────────
...
  /catering/menus:
    get:
      operationId: listCateringMenus
      tags: [catering]
      summary: List catering menu packages
      parameters:
        - in: query
          name: lang
          schema:
            type: string
            default: ru
      responses:
        "200":
          description: Catering menus
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CateringMenu"
    post:
      operationId: createCateringMenu
      tags: [admin, catering]
      summary: Create a catering menu package
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CateringMenuInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CateringMenu"
  /catering/menus/{id}:
    put:
      operationId: updateCateringMenu
      tags: [admin, catering]
      summary: Update a catering menu
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CateringMenuInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CateringMenu"
    delete:
      operationId: deleteCateringMenu
      tags: [admin, catering]
      summary: Delete a catering menu
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /catering/bookings:
    get:
      operationId: listCateringBookings
      tags: [admin, catering]
      summary: List catering bookings (admin)
      parameters:
        - in: query
          name: status
          schema:
            type: ["string", "null"]
            enum: [pending, confirmed, cancelled, null]
      responses:
        "200":
          description: List of catering bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CateringBooking"
    post:
      operationId: createCateringBooking
      tags: [catering]
      summary: Submit a catering booking request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CateringBookingInput"
      responses:
        "201":
          description: Booking submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CateringBooking"
  /catering/bookings/{id}:
    patch:
      operationId: updateCateringBookingStatus
      tags: [admin, catering]
      summary: Update catering booking status (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CateringStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CateringBooking"

  # ──────────────────── REVIEWS ────────────────────
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid credentials
...
  /admin/me:
    get:
      operationId: getAdminMe
      tags: [admin]
      summary: Get current admin session
      responses:
        "200":
          description: Admin session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Not authenticated
components:
  schemas:
...
    GalleryPhoto:
      type: object
      required: [id, type, imageUrl, captionRu, captionIt, captionEn, sortOrder, featured]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [dishes, interior, events, daily]
        imageUrl:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        captionRu:
          type: string
        captionIt:
          type: string
        captionEn:
          type: string
        sortOrder:
          type: integer
        featured:
          type: boolean
        createdAt:
          type: string
    GalleryPhotoInput:
      type: object
      required: [type, imageUrl]
      properties:
        type:
          type: string
          enum: [dishes, interior, events, daily]
        imageUrl:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        captionRu:
          type: string
        captionIt:
          type: string
        captionEn:
          type: string
        sortOrder:
          type: integer
        featured:
          type: boolean

    # ── Events ──
    Event:
      type: object
      required: [id, titleRu, titleIt, titleEn, titleFr, titleZh, descRu, descIt, descEn, descFr, descZh, date, imageUrl]
      properties:
        id:
          type: integer
        titleRu:
          type: string
        titleIt:
          type: string
        titleEn:
          type: string
        titleFr:
          type: string
        titleZh:
          type: string
        descRu:
          type: string
        descIt:
          type: string
        descEn:
          type: string
        descFr:
          type: string
        descZh:
          type: string
        date:
          type: string
        imageUrl:
          type: ["string", "null"]
        published:
          type: boolean
    EventInput:
      type: object
      required: [titleRu, titleIt, titleEn, titleFr, titleZh, descRu, descIt, descEn, descFr, descZh, date]
      properties:
        titleRu:
          type: string
        titleIt:
          type: string
        titleEn:
          type: string
        titleFr:
          type: string
        titleZh:
          type: string
        descRu:
          type: string
        descIt:
          type: string
        descEn:
          type: string
        descFr:
          type: string
        descZh:
          type: string
        date:
          type: string
        imageUrl:
          type: ["string", "null"]
        published:
          type: boolean

    # ── Reservations ──
...
    CateringMenu:
      type: object
      required: [id, nameRu, nameIt, nameEn, nameFr, nameZh, descRu, descIt, descEn, descFr, descZh, pricePerPerson, minGuests, maxGuests, active]
      properties:
        id:
          type: integer
        nameRu:
          type: string
        nameIt:
          type: string
        nameEn:
          type: string
        nameFr:
          type: string
        nameZh:
          type: string
        descRu:
          type: string
        descIt:
          type: string
        descEn:
          type: string
        descFr:
          type: string
        descZh:
          type: string
        pricePerPerson:
          type: number
        minGuests:
          type: integer
        maxGuests:
          type: integer
        imageUrl:
          type: ["string", "null"]
        active:
          type: boolean
    CateringMenuInput:
      type: object
      required: [nameRu, nameIt, nameEn, nameFr, nameZh, descRu, descIt, descEn, descFr, descZh, pricePerPerson, minGuests, maxGuests]
      properties:
        nameRu:
          type: string
        nameIt:
          type: string
        nameEn:
          type: string
        nameFr:
          type: string
        nameZh:
          type: string
        descRu:
          type: string
        descIt:
          type: string
        descEn:
          type: string
        descFr:
          type: string
        descZh:
          type: string
        pricePerPerson:
          type: number
        minGuests:
          type: integer
        maxGuests:
          type: integer
        imageUrl:
          type: ["string", "null"]
        active:
          type: boolean
    CateringBooking:
      type: object
      required: [id, name, phone, eventDate, guests, cateringMenuId, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        eventDate:
          type: string
        eventLocation:
          type: ["string", "null"]
        guests:
          type: integer
        cateringMenuId:
          type: integer
        status:
          type: string
          enum: [pending, confirmed, cancelled]
        comment:
          type: ["string", "null"]
        adminNotes:
          type: ["string", "null"]
        lang:
          type: string
        createdAt:
          type: string
    CateringBookingInput:
      type: object
      required: [name, phone, eventDate, guests, cateringMenuId]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        eventDate:
          type: string
        eventLocation:
          type: ["string", "null"]
        guests:
          type: integer
        cateringMenuId:
          type: integer
        comment:
          type: ["string", "null"]
        lang:
          type: string
    CateringStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, confirmed, cancelled]
        adminNotes:
          type: ["string", "null"]

    # ── Reviews ──
...
    AdminLoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AdminSession:
      type: object
      required: [authenticated, username]
      properties:
        authenticated:
          type: boolean
        username:
          type: string